% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avg_game.R
\name{avg_game}
\alias{avg_game}
\title{Average Game Function}
\usage{
avg_game(year, df = nba)
}
\arguments{
\item{year}{year number}

\item{df}{dataframe Default to nba}
}
\description{
This function calculates the average game for each player in a given year
}
\examples{
avg_game(1950)
}
\keyword{nba}
