% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_num.R
\name{cor_num}
\alias{cor_num}
\title{Correlation matrix for numeric variable}
\usage{
cor_num(year, df = nba)
}
\arguments{
\item{year}{year number}

\item{df}{dataframe Default to nba}
}
\description{
This function returns the correlation matrix for numeric variables in the nba data set
}
\examples{
cor_num(1950)
}
\keyword{nba}
