% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avg_game.R, R/cor_num.R, R/oldest.R
\docType{data}
\name{nba}
\alias{nba}
\title{This is data to be included in my package}
\description{
This is data to be included in my package

This is data to be included in my package

This is data to be included in my package
}
\references{
\url{https://www.kaggle.com/drgilermo/nba-players-stats/data#Seasons_Stats.csv}

\url{https://www.kaggle.com/drgilermo/nba-players-stats/data#Seasons_Stats.csv}

\url{https://www.kaggle.com/drgilermo/nba-players-stats/data#Seasons_Stats.csv}
}
\author{
Chang Qu \email{cq2203@columbia.edu}

Chang Qu \email{cq2203@columbia.edu}

Chang Qu \email{cq2203@columbia.edu}
}
\keyword{data}
