% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oldest.R
\name{oldest}
\alias{oldest}
\title{Oldest Player Function}
\usage{
oldest(year, df = nba)
}
\arguments{
\item{year}{year number}

\item{df}{dataframe Default to nba}
}
\description{
This function returns the oldest player in a given year. If multiple player have the same age
returns the one with the most game played
}
\examples{
oldest(1950)
}
\keyword{nba}
